Refactor base server implementation to support new session and service operations
