+++ pkg/atomix/election/server.go
@@ -37,7 +37,7 @@ func registerServer(server *grpc.Server, protocol node.Protocol) {
 
 func newServer(protocol node.Protocol) api.LeaderElectionServiceServer {
 	return &Server{
-		SessionizedServer: &server.SessionizedServer{
+		Server: &server.Server{
 			Type:     electionType,
 			Protocol: protocol,
 		},
@@ -46,7 +46,7 @@ func newServer(protocol node.Protocol) api.LeaderElectionServiceServer {
 
 // Server is an implementation of LeaderElectionServiceServer for the election primitive
 type Server struct {
-	*server.SessionizedServer
+	*server.Server
 }
 
 // Enter enters a candidate in the election
@@ -309,7 +309,7 @@ func (s *Server) Events(request *api.EventRequest, srv api.LeaderElectionService
 // Create opens a new session
 func (s *Server) Create(ctx context.Context, request *api.CreateRequest) (*api.CreateResponse, error) {
 	log.Tracef("Received CreateRequest %+v", request)
-	header, err := s.OpenSession(ctx, request.Header, request.Timeout)
+	header, err := s.CreateService(ctx, request.Header)
 	if err != nil {
 		return nil, err
 	}
@@ -338,7 +338,7 @@ func (s *Server) KeepAlive(ctx context.Context, request *api.KeepAliveRequest) (
 func (s *Server) Close(ctx context.Context, request *api.CloseRequest) (*api.CloseResponse, error) {
 	log.Tracef("Received CloseRequest %+v", request)
 	if request.Delete {
-		header, err := s.Delete(ctx, request.Header)
+		header, err := s.DeleteService(ctx, request.Header)
 		if err != nil {
 			return nil, err
 		}
@@ -349,7 +349,7 @@ func (s *Server) Close(ctx context.Context, request *api.CloseRequest) (*api.Clo
 		return response, nil
 	}
 
-	header, err := s.CloseSession(ctx, request.Header)
+	header, err := s.CloseService(ctx, request.Header)
 	if err != nil {
 		return nil, err
 	}
